+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.1.0
 
 - Initial release of Paginate
-- Support for Kohana Database, Kohana ORM and Dispatch
+- Support for Kohana Database, Kohana ORM and REST (via Dispatch)
 - Shipped with separate wrapper for DataTables
 - All tests pass "OK (6 tests, 12 assertions)"